feat: refactor security package structure and update dependencies in build.gradle
