+++ .actrc
@@ -1,0 +1,1 @@
+--container-architecture linux/arm64
