+++ .github/workflows/test-secrets-loading.yml
@@ -1,0 +1,29 @@
+# export DOCKER_HOST="unix:///var/folders/x4/rbwfkbt52mg_xvm4gwvmjrpc0000gn/T/podman/podman-machine-default-api.sock"
+# act workflow_dispatch -W .github/workflows/test-secrets-loading.yml --secret-file .secrets
+name: Test Secrets Loading
+
+on:
+  workflow_dispatch:
+
+jobs:
+  test-load-secrets:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout with submodules
+        uses: actions/checkout@v4
+        with:
+          submodules: true
+          token: ${{ secrets.GIT_ACCESS_TOKEN }}
+
+      - name: Load secrets
+        uses: ./.github/actions/load-secrets
+        with:
+          yaml-file: git.environment-variables/sample.yaml
+          prefix: MOCK_
+
+      - name: Test
+        run: |
+          echo "=== 환경변수 ==="
+          env | grep "^MOCK_"
+          echo "Private Key: $MOCK_MOCK_PRIVATE_KEY"
+          echo "Replica Key: $MOCK_MOCK_REPLICA_PRIVATE_KEY"
