+++ bottlenote-mono/src/main/java/app/bottlenote/global/security/CustomUserContext.java
@@ -1,4 +1,4 @@
-package app.bottlenote.global.security.customPrincipal;
+package app.bottlenote.global.security;
 
 import java.util.List;
 import lombok.Getter;
