+++ bottlenote-product-api/build.gradle
@@ -20,10 +20,6 @@ dependencies {
 	implementation libs.spring.boot.starter.web
 	implementation libs.spring.boot.starter.validation
 
-	// Test Dependencies
-	testImplementation libs.spring.security.test
-	testImplementation libs.mockito.inline
-
 	// Test - Database
 	testImplementation libs.spring.boot.starter.data.jpa
 	testImplementation libs.spring.boot.starter.data.redis
@@ -37,9 +33,15 @@ dependencies {
 	annotationProcessor libs.jakarta.annotation.api
 	annotationProcessor libs.jakarta.persistence.api
 
-	// Test - JWT
-	testImplementation libs.bundles.jwt.complete
-	testImplementation libs.google.guava
+	// JWT
+	implementation libs.bundles.jwt.complete
+	implementation libs.google.guava
+
+	// Security
+	implementation libs.spring.boot.starter.security
+	implementation libs.spring.security.test
+	testImplementation libs.spring.security.test
+	testImplementation libs.mockito.inline
 
 	// Test - TestContainers
 	testImplementation libs.bundles.testcontainers.complete
